Move CRUD value types to useBoardData.
